backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
pydantic-settings>=2.0
supabase>=2.0
PyJWT>=2.8
python-multipart>=0.0.9
pdf2image>=1.16
pdfplumber>=0.10
pypdf>=4.0
pytesseract>=0.3
weasyprint>=60.0
